+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/AbstractResultSet.java
@@ -847,7 +847,7 @@ abstract class AbstractResultSet<R> extends AbstractStructReader implements Resu
     public void close(@Nullable String message) {
       if (stream != null) {
         stream.close(message);
-        span.end();
+        span.end(TraceUtil.END_SPAN_OPTIONS);
       }
     }