+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/TraceUtil.java
@@ -21,15 +21,16 @@ import com.google.common.collect.ImmutableMap;
 import com.google.spanner.v1.Transaction;
 import io.opencensus.contrib.grpc.util.StatusConverter;
 import io.opencensus.trace.AttributeValue;
+import io.opencensus.trace.EndSpanOptions;
 import io.opencensus.trace.Span;
 import io.opencensus.trace.Status;
-import io.opencensus.trace.Tracing;
-import io.opencensus.trace.export.SampledSpanStore;
-import java.util.Arrays;
 import java.util.Map;
 
 /** Utility methods for tracing. */
 class TraceUtil {
+
+  static final EndSpanOptions END_SPAN_OPTIONS =
+      EndSpanOptions.builder().setSampleToLocalSpanStore(true).build();
 
   static Map<String, AttributeValue> getTransactionAnnotations(Transaction t) {
     return ImmutableMap.of(
@@ -58,7 +59,7 @@ class TraceUtil {
       endSpanWithFailure(span, (SpannerException) e);
     } else {
       span.setStatus(Status.INTERNAL.withDescription(e.getMessage()));
-      span.end();
+      span.end(END_SPAN_OPTIONS);
     }
   }
 
@@ -66,13 +67,6 @@ class TraceUtil {
     span.setStatus(
         StatusConverter.fromGrpcStatus(e.getErrorCode().getGrpcStatus())
             .withDescription(e.getMessage()));
-    span.end();
-  }
-
-  static void exportSpans(String... spans) {
-    SampledSpanStore store = Tracing.getExportComponent().getSampledSpanStore();
-    if (store != null) {
-      store.registerSpanNamesForCollection(Arrays.asList(spans));
-    }
+    span.end(END_SPAN_OPTIONS);
   }
 }